requirements.txt
wheel
python-dotenv
astral
pytz
opencv-python